requirements.txt
Flask==0.10.1
gnureadline==6.3.3
gunicorn==19.1.1
ipython==2.3.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pip-tools==0.3.5
PyJWT==0.3.1
requests==2.5.0
Werkzeug==0.9.6
